Replace Gemini with Dify
